fixup! https://github.com/OpenPaaS-Suite/esn/issues/6: Added 2 new components, esn-app-grid & esn-popover
