+++ src/components/esn-app-grid/esn-app-grid.tsx
@@ -1,22 +1,7 @@
 import { Component, Prop, ComponentInterface, State, Host, h } from '@stencil/core';
 import { Application } from './esn-app-grid.types';
-import { AppGridTogglerIcon } from './icons/AppGridIcon';
-import { CalendarAppIcon } from './icons/CalendarAppIcon';
-import { InboxAppIcon } from './icons/InboxAppIcon';
-import { AdminAppIcon } from './icons/AdminAppIcon';
-import { ContactsAppIcon } from './icons/ContactsAppIcon';
-import { LinShareAppIcon } from './icons/LinShareAppIcon';
-import { UnknownAppIcon } from './icons/UnknownAppIcon';
-
-const iconMapping = {
-  Admin: AdminAppIcon,
-  Calendar: CalendarAppIcon,
-  Contacts: ContactsAppIcon,
-  Inbox: InboxAppIcon,
-  LinShare: LinShareAppIcon
-};
-
-const isIconAvailable = (appName: string) => Object.keys(iconMapping).includes(appName);
+import { AppGridTogglerIcon } from '../../icons/AppGridIcon';
+import { getAppIcon } from '../../utils/app-icons';
 
 @Component({
   tag: 'esn-app-grid',
@@ -41,22 +26,11 @@ export class EsnAppGrid implements ComponentInterface {
 
       const applications = JSON.parse(this.serializedApplications);
 
-      if (Array.isArray(applications) && applications.length) {
-        this.applications = applications.map(({ name, url }) => {
-          if (!isIconAvailable(name)) {
-            console.warn(
-              `The application with name ${name} does not have an icon available. A fallback icon will be used instead.`
-            );
-          }
-
-          return {
-            name,
-            url
-          };
-        });
-      } else {
+      if (!Array.isArray(applications) || !applications.length) {
         throw new Error("The parsed 'serializedApplications' has an invalid JSON structure");
       }
+
+      this.applications = applications;
     } catch (err) {
       console.error('Something went wrong', err);
       this.applications = [];
@@ -78,18 +52,16 @@ export class EsnAppGrid implements ComponentInterface {
           </button>
           <esn-popover ref={el => (this.popover = el as HTMLElement)}>
             <div class="esn-app-grid__popover-content">
-              {this.applications.map(application => {
-                const IconComponent = isIconAvailable(application.name)
-                  ? iconMapping[application.name]
-                  : UnknownAppIcon;
+              {this.applications.map(({ name, url }) => {
+                const IconComponent = getAppIcon(name);
 
                 return (
                   <div class="esn-app-grid__app-item">
-                    <a href={application.url}>
+                    <a href={url}>
                       <div class="esn-app-grid__app-icon">
                         <IconComponent />
                       </div>
-                      <span class="label">{application.name}</span>
+                      <span class="label">{name}</span>
                     </a>
                   </div>
                 );
