+++ src/icons/AdminAppIcon.tsx
