+++ src/icons/InboxAppIcon.tsx
