+++ src/utils/app-icons.ts
@@ -1,0 +1,22 @@
+import { AdminAppIcon } from '../icons/AdminAppIcon';
+import { CalendarAppIcon } from '../icons/CalendarAppIcon';
+import { ContactsAppIcon } from '../icons/ContactsAppIcon';
+import { InboxAppIcon } from '../icons/InboxAppIcon';
+import { LinShareAppIcon } from '../icons/LinShareAppIcon';
+import { UnknownAppIcon } from '../icons/UnknownAppIcon';
+
+const iconMapping = {
+  Admin: AdminAppIcon,
+  Calendar: CalendarAppIcon,
+  Contacts: ContactsAppIcon,
+  Inbox: InboxAppIcon,
+  LinShare: LinShareAppIcon
+};
+
+export const getAppIcon = (appName: string) => {
+  if (iconMapping[appName]) return iconMapping[appName];
+
+  console.warn(`The application with name ${appName} does not have an icon available. A fallback icon will be used instead.`);
+
+  return UnknownAppIcon;
+}
